Style changes, accordion selected highlighted
