+++ frontend/app/(auth)/login.jsx
@@ -13,6 +13,8 @@ import { signIn } from "./authService";
 import { useGoogleSignIn } from "./authService"; // ✅ Import fixed Google Sign-In Hook
 import TextField from "@/components/common/Textfield.jsx";
 import TextButton from "@/components/common/TextButton.jsx";
+import KeyboardDismissWrapper from "@/components/util/KeyboardWrapper.jsx";
+
 
 export default function LoginPage() {
   const router = useRouter();
