+++ frontend/package.json
@@ -55,7 +55,7 @@
     "react-native-screens": "~4.4.0",
     "react-native-toast-message": "^2.2.1",
     "react-native-view-shot": "~4.0.3",
-    "react-native-web": "~0.19.6",
+    "react-native-web": "~0.19.13",
     "react-native-webview": "13.12.5",
     "react-redux": "^9.2.0",
     "undefined": "expo-auth-session/providers/google"
